서버 Long 타입 변수 수정 및 프론트 자동 로그인기능 수정

+++ backend/src/main/java/com/chamchi/backend/domain/category/CategoryCount.java
@@ -14,11 +14,11 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CategoryCount {
     @Id
-    private Long id;
+    private Integer id;
     private String name;
-    private Long count;
+    private Integer count;
 
-    public CategoryCount(Long id, String name, Long count) {
+    public CategoryCount(Integer id, String name, Integer count) {
         this.id = id;
         this.name = name;
         this.count = count;
