+++ backend/src/main/java/com/chamchi/backend/repository/post/image/ImageTempRepository.java
@@ -3,6 +3,6 @@ package com.chamchi.backend.repository.post.image;
 import com.chamchi.backend.domain.post.image.ImageTemp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageTempRepository extends JpaRepository<ImageTemp, Long> {
+public interface ImageTempRepository extends JpaRepository<ImageTemp, Integer> {
     ImageTemp findByImageName(String imageName);
 }
